requirements.txt
django==2.2.20
pillow==8.2.0
